+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v7.1.4
+----------
+ * Remove task to retry errored messages which now handled in mailroom
+
+v7.1.2
+----------
+ * Update poetry dependencies
+ * Update to latest editor
+
 v7.1.1
 ----------
  * Remove channel alert notifications as these will become incidents